async improvements

+++ shaders/Voxel_Rendering_Fragment.glsl
@@ -1,7 +1,7 @@
-precision highp float;
-precision highp int;
-precision highp sampler2D;
-precision highp sampler3D;
+precision mediump float;
+precision mediump int;
+precision mediump sampler2D;
+precision mediump sampler3D;
 
 uniform float uBlurRatio;
